Add a PSP-specific Makefile.
Convert doubles to floats.

Requires SDL + OpenGL which hasn't been submitted yet.

+++ rRootage/src/barragemanager.h
@@ -18,7 +18,7 @@
 
 typedef struct {
   BulletMLParser *bulletml;
-  double maxRank, rank;
+  float maxRank, rank;
 } Barrage;
 
 #define BARRAGE_TYPE_NUM 6
